Switch to OrderedDict for SET and SEQUENCE types
When working with some systems/standards, the order of the elements in a given
constructed type must be maintained, as processing at the receiver of
such data is required to happen in order.  One example of such formats
is the SimAlliace / TCA Interoperable Profile Package for eSIM.

As normal python 'dict' does not preserve the order, let's switch to
OrderedDict.